Changes in CKKS Encoder from int to ulong

+++ net/dll/seal/batchencoder.cpp
@@ -207,12 +207,12 @@ SEALDLL HRESULT SEALCALL BatchEncoder_Decode3(void* thisptr, void* plain, void* 
     }
 }
 
-SEALDLL HRESULT SEALCALL BatchEncoder_GetSlotCount(void* thisptr, int* slot_count)
+SEALDLL HRESULT SEALCALL BatchEncoder_GetSlotCount(void* thisptr, uint64_t* slot_count)
 {
     BatchEncoder* encoder = FromVoid<BatchEncoder>(thisptr);
     IfNullRet(encoder, E_POINTER);
     IfNullRet(slot_count, E_POINTER);
 
-    *slot_count = static_cast<int>(encoder->slot_count());
+    *slot_count = encoder->slot_count();
     return S_OK;
 }
